# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReconFlow AI",
    "description": "Delivery-platform reconciliation for restaurants.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/uploaded-files/{uploaded_file_id}/normalization": {
      "post": {
        "tags": [
          "data-normalization"
        ],
        "summary": "Normalize Uploaded File",
        "operationId": "normalize_uploaded_file_uploaded_files__uploaded_file_id__normalization_post",
        "parameters": [
          {
            "name": "uploaded_file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uploaded File Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalizationRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/matching": {
      "post": {
        "tags": [
          "matching-engine"
        ],
        "summary": "Run Matching",
        "operationId": "run_matching_analyses__analysis_id__matching_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchingRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/matches/{reconciliation_match_id}/review": {
      "post": {
        "tags": [
          "manual-review"
        ],
        "summary": "Review Match",
        "operationId": "review_match_matches__reconciliation_match_id__review_post",
        "parameters": [
          {
            "name": "reconciliation_match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconciliation Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/discrepancies/{discrepancy_id}/review": {
      "post": {
        "tags": [
          "manual-review"
        ],
        "summary": "Review Discrepancy",
        "operationId": "review_discrepancy_discrepancies__discrepancy_id__review_post",
        "parameters": [
          {
            "name": "discrepancy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Discrepancy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscrepancyReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscrepancyReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}/manual-match": {
      "post": {
        "tags": [
          "manual-review"
        ],
        "summary": "Create Manual Match",
        "operationId": "create_manual_match_analyses__analysis_id__manual_match_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualMatchRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualMatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/analyses/{analysis_id}/notifications": {
      "post": {
        "tags": [
          "notification"
        ],
        "summary": "Send Notification",
        "operationId": "send_notification_analyses__analysis_id__notifications_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscrepancyReviewDecision": {
        "type": "string",
        "enum": [
          "ACKNOWLEDGE",
          "DISPUTE"
        ],
        "title": "DiscrepancyReviewDecision"
      },
      "DiscrepancyReviewRequest": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/DiscrepancyReviewDecision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DiscrepancyReviewRequest"
      },
      "DiscrepancyReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "discrepancy_id": {
            "type": "string",
            "title": "Discrepancy Id"
          },
          "decision": {
            "$ref": "#/components/schemas/DiscrepancyReviewDecision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewed_by": {
            "type": "string",
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "discrepancy_id",
          "decision",
          "note",
          "reviewed_by",
          "reviewed_at"
        ],
        "title": "DiscrepancyReviewResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualMatchRequest": {
        "properties": {
          "pos_transaction_id": {
            "type": "string",
            "title": "Pos Transaction Id"
          },
          "platform_transaction_id": {
            "type": "string",
            "title": "Platform Transaction Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "pos_transaction_id",
          "platform_transaction_id"
        ],
        "title": "ManualMatchRequest"
      },
      "ManualMatchResponse": {
        "properties": {
          "reconciliation_match_id": {
            "type": "string",
            "title": "Reconciliation Match Id"
          },
          "pos_transaction_id": {
            "type": "string",
            "title": "Pos Transaction Id"
          },
          "platform_transaction_id": {
            "type": "string",
            "title": "Platform Transaction Id"
          },
          "review": {
            "$ref": "#/components/schemas/MatchReviewResponse"
          }
        },
        "type": "object",
        "required": [
          "reconciliation_match_id",
          "pos_transaction_id",
          "platform_transaction_id",
          "review"
        ],
        "title": "ManualMatchResponse"
      },
...
      "MatchReviewDecision": {
        "type": "string",
        "enum": [
          "CONFIRM",
          "REJECT",
          "MANUALLY_PAIRED"
        ],
        "title": "MatchReviewDecision"
      },
      "MatchReviewRequest": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/MatchReviewDecision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "MatchReviewRequest"
      },
      "MatchReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "reconciliation_match_id": {
            "type": "string",
            "title": "Reconciliation Match Id"
          },
          "decision": {
            "$ref": "#/components/schemas/MatchReviewDecision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewed_by": {
            "type": "string",
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reviewed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "reconciliation_match_id",
          "decision",
          "note",
          "reviewed_by",
          "reviewed_at"
        ],
        "title": "MatchReviewResponse"
      },
      "MatchingRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "status": {
            "$ref": "#/components/schemas/MatchingStatus"
          },
          "matched_count": {
            "type": "integer",
            "title": "Matched Count"
          },
          "unmatched_pos_count": {
            "type": "integer",
            "title": "Unmatched Pos Count"
          },
          "unmatched_platform_count": {
            "type": "integer",
            "title": "Unmatched Platform Count"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_id",
          "status",
          "matched_count",
          "unmatched_pos_count",
          "unmatched_platform_count",
          "checked_at"
        ],
        "title": "MatchingRunResponse"
      },
      "MatchingStatus": {
        "type": "string",
        "enum": [
          "COMPLETED",
          "FAILED"
        ],
        "title": "MatchingStatus"
      },
      "NormalizationRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "uploaded_file_id": {
            "type": "string",
            "title": "Uploaded File Id"
          },
          "status": {
            "$ref": "#/components/schemas/NormalizationStatus"
          },
          "rows_created": {
            "type": "integer",
            "title": "Rows Created"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/NormalizationWarningResponse"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "uploaded_file_id",
          "status",
          "rows_created",
          "checked_at",
          "warnings"
        ],
        "title": "NormalizationRunResponse"
      },
      "NormalizationStatus": {
        "type": "string",
        "enum": [
          "COMPLETED",
          "FAILED"
        ],
        "title": "NormalizationStatus"
      },
      "NormalizationWarningResponse": {
        "properties": {
          "row_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Number"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "row_number",
          "field",
          "message"
        ],
        "title": "NormalizationWarningResponse"
      },
      "NotificationEventType": {
        "type": "string",
        "enum": [
          "ANALYSIS_COMPLETED",
          "ANALYSIS_FAILED"
        ],
        "title": "NotificationEventType"
      },
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/NotificationEventType"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time",
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_id",
          "event_type",
          "channel",
          "recipient",
          "status",
          "failure_reason",
          "sent_at"
        ],
        "title": "NotificationResponse"
      },
      "NotificationStatus": {
        "type": "string",
        "enum": [
          "SENT",
          "FAILED"
        ],
        "title": "NotificationStatus"
      },
...
      "SendNotificationRequest": {
        "properties": {
          "event_type": {
            "$ref": "#/components/schemas/NotificationEventType"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "recipient"
        ],
        "title": "SendNotificationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}